# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "GkEntity": {
        "description": "P1/P2 \u8fd4\u56de\u4f53\u3002relations \u662f {\u5173\u7cfb\u540d: [\u5bf9\u65b9\u77ed\u7801]} \u7684\u888b\u5b50\uff08M7 \u8bed\u4e49\uff09\u3002",
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "properties": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Properties"
          },
          "relations": {
            "anyOf": [
              {
                "additionalProperties": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relations"
          },
          "versions": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/GkVersion"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Versions"
          }
        },
        "title": "GkEntity",
        "type": "object"
      },
      "GkEntitySummary": {
        "description": "\u53ea\u8bfb\u951a\u70b9\u76ee\u5f55\u4e2d\u7684\u6700\u5c0f\u5b9e\u4f53\u5f62\u72b6\u3002",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "code",
          "label"
        ],
        "title": "GkEntitySummary",
        "type": "object"
      },
...
      "GkVersion": {
        "description": "P2 \u7248\u672c\u94fe\u4e2d\u7684\u4e00\u7248\u3002\u5168\u90e8\u53ef\u7a7a\u2014\u2014\u670d\u52a1\u7aef\u7f3a\u5b57\u6bb5\u4e0d\u5d29\uff0c\u6d88\u8d39\u65b9\u6309\u7f3a\u7701\u5904\u7406\u3002",
        "properties": {
          "at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "At"
          },
          "index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Index"
          },
          "meetingLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meetinglabel"
          },
          "statement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Statement"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "GkVersion",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBasic": {
        "scheme": "basic",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Standalone governed Memory Service with a native API and a temporary Clark GraphKnowledge compatibility fa\u00e7ade.",
    "title": "TKOS Memory Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/entities": {
      "get": {
        "description": "List exact-title ISS anchor candidates for the clark compatibility path.",
        "operationId": "entities_api_v1_entities_get",
        "parameters": [
          {
            "in": "query",
            "name": "prefix",
            "required": true,
            "schema": {
              "const": "ISS-",
              "title": "Prefix",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GkEntitySummary"
                  },
                  "title": "Response Entities Api V1 Entities Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Entities",
        "tags": [
          "entities"
        ]
      }
    },
    "/api/v1/entities/{code}": {
      "get": {
        "description": "Return one scoped WM entity in M7 P1/P2 shape.",
        "operationId": "entity_api_v1_entities__code__get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "view",
            "required": false,
            "schema": {
              "default": "latest",
              "enum": [
                "latest",
                "chain"
              ],
              "title": "View",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GkEntity"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Entity",
        "tags": [
          "entities"
        ]
      }
    },
...
  }
}